cloud-engineer-task-temmy-master/kittehnet/requirements.txt
flask==1.1.2
mysql-connector==2.2.9
